modified architecture and added wavelet
