+++ model_weights/baseline_seg/checkpoint
@@ -1,0 +1,2 @@
+model_checkpoint_path: "weights"
+all_model_checkpoint_paths: "weights"
